+++ CHANGELOG.md
@@ -4,7 +4,10 @@ LR(0) parser Change Log
 1.6.0 (next)
 -----
 
-No changes yet.
+*   Deprecated: `\VovanVE\parser\lexer\ParseException` now replaced with
+    `\VovanVE\parser\errors\UnknownCharacterException`.
+*   End: `\VovanVE\parser\SyntaxException` divided into more detailed exceptions
+    to be explained to end-user easily.
 
 1.5.3
 -----